separate resolver logic from lookup via api
