+++ packages/orn-locator/package.json
@@ -6,11 +6,15 @@
   "types": "index.d.ts",
   "exports": {
     ".": {
-      "node": "./dist/nodejs.js",
-      "browser": "./dist/browser.js",
+      "node": "./dist/index.nodejs.js",
+      "browser": "./dist/index.browser.js",
       "default": "./dist/index.js"
     },
-    "./": "./dist/"
+    "./resolve": {
+      "node": "./dist/resolve.nodejs.js",
+      "browser": "./dist/resolve.browser.js",
+      "default": "./dist/resolve.js"
+    }
   },
   "files": [
     "dist"
@@ -29,7 +33,10 @@
     "coverage-report": "open coverage/index.html",
     "test": "yarn jest",
     "build:example": "yarn build && yarn webpack -c ./example/webpack.config.js",
+    "start": "./script/build.bash --watch",
+    "clean": "rm -rf dist build",
     "build": "./script/build.bash",
+    "build:clean": "yarn clean && yarn build",
     "ci:build": "yarn build --clean",
     "ci:test": "yarn jest --coverage",
     "ci:typecheck": "tsc --noEmit",
