+++ packages/orn-locator/src/index.browser.ts
