+++ packages/orn-locator/src/index.nodejs.ts
@@ -1,0 +1,7 @@
+import nodeFetch from 'node-fetch';
+
+declare var globalThis: any;
+
+globalThis.fetch = nodeFetch;
+
+export * from './';
