+++ packages/orn-locator/src/resolve.browser.ts
@@ -1,0 +1,1 @@
+export * from './resolve';
